Add content to the repository
Add project "python-tuf" with two products
 * "python-tuf" with releases 1.0.0 and 1.1.0
 * "src" with releases 1.0.0 and 1.1.0

There is an index.json decribing these releases but the target paths are
also well formed and fully describe the content
(targetpath = project/product/version/artifact ). This would enable the
client to just look at the metadata to find the projects/ products
instead of parsing index.json files

This was produced by copying the content to the correct place, then
running for each artifact:

  ARTIFACT="python-tuf/python-tuf/1.0.0/tuf-1.0.0-py3-none-any.whl"
  tufrepo edit targets add-target "$ARTIFACT" "../content/$ARTIFACT"

This modifies the metadata and creates the symlink.

Finally, snapshot and timestamp are updated:

  tufrepo snapshot